scripts/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pocketbase-async>=0.12.0",
    "pydantic-ai-slim[groq,mcp,openai]>=1.11.0",
    "tiktoken>=0.12.0",
]
